Allow for nested objects in the configuration files, and move the configuration specific file reading things to their own file.
